Two conflicts, both from #716 (remove the two constant-bit tools):

tools/test_constantbitprop/test_cbitp.cpp -- modify/delete. Took the
deletion. The tool is gone upstream, so the ownership fix this branch
made to its main() is moot.

tools/measure_constantbitprop/measure.cpp -- content. #716's version is
the base: its include set (Relations.h moved into the tool directory, the
transfer-function headers that replaced the deleted Functions.cpp) and
its file body are kept intact, with this branch's main() rewrite applied
on top and the STPManager/STP.h include added alongside theirs.

#716 did not touch measure.cpp's main(), so the ordering hazard this
branch exists to fix is still present on master: `delete mgr` still runs
inside the Cpp_interface's scope. measure.cpp is now the only one-argument
Cpp_interface(STPMgr&) caller left in the tree.

+++ include/stp/c_interface.h
@@ -1267,7 +1267,7 @@ DLL_PUBLIC const char* exprName(Expr e);
 
 //! \brief Returns the internal node ID of the given expression.
 //!
-DLL_PUBLIC int getExprID(Expr ex);
+DLL_PUBLIC uint64_t getExprID(Expr ex);
 
 //! \brief Parses the given string in CVC or SMTLib1.0 format and extracts
 //!        query and assertion information into the 'outQuery' and 'outAsserts'
